Add files via upload
Added Java files and accompanying images for the quiz game. 
i) Around.java
Contains the implementation of the question and choice display, checking answer correctness and calculating scores. 
Other features:
. Image display along with questions
. Timer countdown
. Game termination notification and score display
ii) AroundCanvas.java
Java file contains the graphics function that paints within the Applet to show the quiz-like game display. It contains a panel for the question, picture and choice buttons. The implementation on a Java Applet was to give the game a retro feel. 

Images uploaded along with the files are the accompany the questions. 

Other details:
Each game consists of a round of 10 questions. 
The player has 15 seconds to answer each question. 
Please have Java Applets enabled on your PC to run this game.